pyproject.toml
[tool.poetry]
name = "technical test"
version = "0.1.1"
description = "Technical test"
authors = ["Pablo Burgos <hispar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.2"
django-model-utils = "^4.3.1"
django-ninja = "^0.22.2"
django-cors-headers = "^4.1.0"
pytz = "^2023.3"
streamlit = "^1.23.1"
pandas = "^2.0.2"
numpy = "^1.25.0"
black = "^23.3.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
